pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "aas-holo-shard"
version = "0.1.0"
description = "Threshold cryptography utilities for securing AASX packages using Shamir's Secret Sharing."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "AAS-Holo-Shard Contributors" }
]
keywords = [
  "aas",
  "aasx",
  "shamir",
  "secret-sharing",
  "threshold-cryptography",
  "digital-twin",
  "industry-4.0",
  "security"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security :: Cryptography"
]
dependencies = [
  "pycryptodome>=3.19"
]

[project.optional-dependencies]
dev = [
  "black>=24.4.2",
  "bandit>=1.7.7",
  "pre-commit>=3.7.1",
  "ruff>=0.5.0"
]
test = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "hypothesis>=6.100.0"
]

[tool.black]
line-length = 100

target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["aas_holo_shard"]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.hatch.build.targets.wheel]
packages = ["src/aas_holo_shard"]
